move DeleteConfirmationDialogs into context
