+++ apps/studio/components/interfaces/TableGridEditor/TableGridEditor.tsx
@@ -25,15 +25,19 @@ import { Admonition, GenericSkeletonLoader } from 'ui-patterns'
 import { useIsTableEditorTabsEnabled } from '../App/FeaturePreview/FeaturePreviewContext'
 import SidePanelEditor from './SidePanelEditor/SidePanelEditor'
 import TableDefinition from './TableDefinition'
+import DeleteConfirmationDialogs from './DeleteConfirmationDialogs'
+import { TablesData } from 'data/tables/tables-query'
 
 export interface TableGridEditorProps {
   isLoadingSelectedTable?: boolean
   selectedTable?: Entity
+  onAfterDeleteTable?: (tables: TablesData) => void
 }
 
 export const TableGridEditor = ({
   isLoadingSelectedTable = false,
   selectedTable,
+  onAfterDeleteTable,
 }: TableGridEditorProps) => {
   const router = useRouter()
   const editor = useEditorType()
@@ -161,6 +165,7 @@ export const TableGridEditor = ({
           selectedTable={isTableLike(selectedTable) ? selectedTable : undefined}
           onTableCreated={onTableCreated}
         />
+        <DeleteConfirmationDialogs selectedTable={selectedTable} />
       </TableEditorTableStateContextProvider>
     </div>
   )
