+++ apps/studio/pages/project/[ref]/editor/[id].tsx
@@ -1,9 +1,7 @@
-import { useRouter } from 'next/router'
-import { useCallback, useEffect } from 'react'
+import { useEffect } from 'react'
 
 import { useParams } from 'common'
 import { useIsTableEditorTabsEnabled } from 'components/interfaces/App/FeaturePreview/FeaturePreviewContext'
-import DeleteConfirmationDialogs from 'components/interfaces/TableGridEditor/DeleteConfirmationDialogs'
 import { TableGridEditor } from 'components/interfaces/TableGridEditor/TableGridEditor'
 import DefaultLayout from 'components/layouts/DefaultLayout'
 import { EditorBaseLayout } from 'components/layouts/editors/EditorBaseLayout'
@@ -11,7 +9,6 @@ import { useProjectContext } from 'components/layouts/ProjectLayout/ProjectConte
 import TableEditorLayout from 'components/layouts/TableEditorLayout/TableEditorLayout'
 import TableEditorMenu from 'components/layouts/TableEditorLayout/TableEditorMenu'
 import { useTableEditorQuery } from 'data/table-editor/table-editor-query'
-import { TablesData } from 'data/tables/tables-query'
 import { addTab, createTabId, getTabsStore } from 'state/tabs'
 import type { NextPageWithLayout } from 'types'
 
@@ -26,19 +23,6 @@ const TableEditorPage: NextPageWithLayout = () => {
     connectionString: project?.connectionString,
     id,
   })
-
-  const router = useRouter()
-  const onAfterDeleteTable = useCallback(
-    (tables: TablesData) => {
-      // For simplicity for now, we just open the first table within the same schema
-      if (tables.length > 0) {
-        router.push(`/project/${projectRef}/editor/${tables[0].id}`)
-      } else {
-        router.push(`/project/${projectRef}/editor`)
-      }
-    },
-    [router, projectRef]
-  )
 
   /**
    * Effect: Creates or updates tab when table is loaded
@@ -73,10 +57,6 @@ const TableEditorPage: NextPageWithLayout = () => {
   return (
     <>
       <TableGridEditor isLoadingSelectedTable={isLoading} selectedTable={selectedTable} />
-      <DeleteConfirmationDialogs
-        selectedTable={selectedTable}
-        onAfterDeleteTable={onAfterDeleteTable}
-      />
     </>
   )
 }
